Metrics: Wire up `objstore_bucket_operation_duration_seconds` in `iter` operations
Trying to get some statistics in Mimir about the time we're spending in our `iter` operations I realised that despite these operations showing up as part of the labels their value was always 0.

Soon enough and with the help of Peter, I realised that these were not wired up in `objstore.go`.

This PR does exactly that and measures the duration of the whole `iter` operation. There's one big caveat though, `iter` operations take a callback as an argument which is executed per each item returned. The computed duration will include whatever time we spend on each of these callbacks which might an undesirable effect for users of the library.

Sadly, I can't get around this without changing the interface for `iter` to (e.g.) return JUST the time the network operation took. In the meantime, I've done the best I've could without incurring in a big change and documented the behaviour as part of the Histogram's HELP.

Co-Authored-By: Peter Štibraný <895919+pstibrany@users.noreply.github.com>
Signed-off-by: gotjosh <josue.abreu@gmail.com>