download: dont cleanup directory after download failed
DownloadDir has the option to ignore paths that we have already on disk.
In theory that makes downloads resumable, but in practice this doesnt
work because we delete everything if one download fails. This means if
one chunk in 1TiB block fails to download for whatever reason we start
all over.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>